@@ -30,17 +30,16 @@
     "test:karma": "karma start --single-run"
   },
   "dependencies": {
-    "@angular/animations": "^7.0.3",
-    "@angular/common": "^7.0.3",
-    "@angular/compiler": "^7.0.3",
-    "@angular/compiler-cli": "^7.0.3",
-    "@angular/core": "^7.0.3",
-    "@angular/forms": "^7.0.3",
-    "@angular/http": "^7.0.3",
-    "@angular/platform-browser": "^7.0.3",
-    "@angular/platform-browser-dynamic": "^7.0.3",
-    "@angular/router": "^7.0.3",
-    "@angular/upgrade": "^7.0.3",
+    "@angular/animations": "~8.1.2",
+    "@angular/common": "~8.1.2",
+    "@angular/compiler": "~8.1.2",
+    "@angular/compiler-cli": "~8.1.2",
+    "@angular/core": "~8.1.2",
+    "@angular/forms": "~8.1.2",
+    "@angular/platform-browser": "~8.1.2",
+    "@angular/platform-browser-dynamic": "~8.1.2",
+    "@angular/router": "~8.1.2",
+    "@angular/upgrade": "~8.1.2",
     "angular": "~1.3.2",
     "angular-split": "1.0.0-rc.3",
     "angular-tree-component": "^8.0.0",
@@ -63,19 +62,19 @@
     "ngx-mask": "^2.5.3",
     "prismjs": "^1.6.0",
     "resize-observer-polyfill": "^1.5.0",
-    "rxjs": "^6.2.2",
-    "rxjs-compat": "^6.2.2",
+    "rxjs": "~6.5.2",
+    "rxjs-compat": "~6.5.2",
     "semver": "5.6.0",
     "source-sans-pro": "^2.0.10",
-    "zone.js": "0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.10.6",
-    "@angular-devkit/build-optimizer": "^0.10.5",
+    "@angular-devkit/build-angular": "~0.801.2",
+    "@angular-devkit/build-optimizer": "~0.801.2",
     "@angular/cdk": "^6.4.7",
     "@babel/core": "^7.1.5",
     "@babel/preset-env": "^7.1.5",
-    "@ngtools/webpack": "7.0.5",
+    "@ngtools/webpack": "^8.1.2",
     "@types/angular": "1.6.6",
     "@types/angular-ui-router": "1.1.40",
     "@types/chance": "0.7.35",
@@ -99,6 +98,7 @@
     "cheerio": "^1.0.0-rc.2",
     "chromedriver": "2.36.0",
     "circular-dependency-plugin": "^5.0.2",
+    "codelyzer": "^5.0.1",
     "copy-webpack-plugin": "4.6.0",
     "css-loader": "1.0.0",
     "file-loader": "1.1.11",
@@ -155,8 +155,8 @@
     "to-string-loader": "1.1.5",
     "ts-node": "^7.0.1",
     "tsickle": "^0.33.1",
-    "tslint": "5.8.0",
-    "typescript": "~3.1.1",
+    "tslint": "^5.11.0",
+    "typescript": "~3.4.5",
     "uglify-js": "3.4.5",
     "webpack": "4.25.1",
     "webpack-cli": "^3.1.2",
