@@ -12,19 +12,19 @@ const ROUTES: Routes = [
         component: ChartsPageComponent,
         children: [
             { path: '', pathMatch: 'full', redirectTo: 'bar-charts' },
-            { path: 'bar-charts', loadChildren: './charts-sections/bar-charts/bar-charts.module#ChartsBarChartsModule' },
-            { path: 'peity-charts', loadChildren: './charts-sections/peity-charts/peity-charts.module#ChartsPeityChartsModule' },
-            { path: 'donut-charts', loadChildren: './charts-sections/donut-charts/donut-charts.module#ChartsDonutChartsModule' },
-            { path: 'flot-options', loadChildren: './charts-sections/flot-options/flot-options.module#ChartsFlotOptionsModule' },
-            { path: 'line-charts', loadChildren: './charts-sections/line-charts/line-charts.module#ChartsLineChartsModule' },
-            { path: 'live-chart', loadChildren: './charts-sections/live-charts/live-charts.module#ChartsLiveChartsModule' },
-            { path: 'organization-chart', loadChildren: './charts-sections/organization-chart/organization-chart.module#ChartsOrganizationChartModule' },
-            { path: 'partition-map', loadChildren: './charts-sections/partition-map/partition-map.module#ChartsPartitionMapModule' },
-            { path: 'sankey-chart', loadChildren: './charts-sections/sankey-chart/sankey-charts.module#ChartsSankeyChartModule' },
-            { path: 'scrollable-chart', loadChildren: './charts-sections/scrollable-chart/scrollable-chart.module#ChartsScrollableChartModule' },
-            { path: 'social-chart', loadChildren: './charts-sections/social-chart/social-chart.module#ChartsSocialChartModule' },
-            { path: 'spark-charts', loadChildren: './charts-sections/spark-charts/spark-charts.module#ChartsSparkChartsModule' },
-            { path: 'timeline-chart', loadChildren: './charts-sections/timeline-chart/timeline-chart.module#ChartsTimelineChartModule' }
+            { path: 'bar-charts', loadChildren: () => import('./charts-sections/bar-charts/bar-charts.module').then(m => m.ChartsBarChartsModule) },
+            { path: 'peity-charts', loadChildren: () => import('./charts-sections/peity-charts/peity-charts.module').then(m => m.ChartsPeityChartsModule) },
+            { path: 'donut-charts', loadChildren: () => import('./charts-sections/donut-charts/donut-charts.module').then(m => m.ChartsDonutChartsModule) },
+            { path: 'flot-options', loadChildren: () => import('./charts-sections/flot-options/flot-options.module').then(m => m.ChartsFlotOptionsModule) },
+            { path: 'line-charts', loadChildren: () => import('./charts-sections/line-charts/line-charts.module').then(m => m.ChartsLineChartsModule) },
+            { path: 'live-chart', loadChildren: () => import('./charts-sections/live-charts/live-charts.module').then(m => m.ChartsLiveChartsModule) },
+            { path: 'organization-chart', loadChildren: () => import('./charts-sections/organization-chart/organization-chart.module').then(m => m.ChartsOrganizationChartModule) },
+            { path: 'partition-map', loadChildren: () => import('./charts-sections/partition-map/partition-map.module').then(m => m.ChartsPartitionMapModule) },
+            { path: 'sankey-chart', loadChildren: () => import('./charts-sections/sankey-chart/sankey-charts.module').then(m => m.ChartsSankeyChartModule) },
+            { path: 'scrollable-chart', loadChildren: () => import('./charts-sections/scrollable-chart/scrollable-chart.module').then(m => m.ChartsScrollableChartModule) },
+            { path: 'social-chart', loadChildren: () => import('./charts-sections/social-chart/social-chart.module').then(m => m.ChartsSocialChartModule) },
+            { path: 'spark-charts', loadChildren: () => import('./charts-sections/spark-charts/spark-charts.module').then(m => m.ChartsSparkChartsModule) },
+            { path: 'timeline-chart', loadChildren: () => import('./charts-sections/timeline-chart/timeline-chart.module').then(m => m.ChartsTimelineChartModule) }
         ],
     },
     {
