@@ -12,7 +12,7 @@ import { CardTabContentDirective } from './card-tab-content.directive';
 export class CardTabComponent implements OnDestroy {
 
     active$: Observable<boolean> = this._tabService.tab$.pipe(map(tab => tab === this));
-    @ContentChild(CardTabContentDirective, { read: TemplateRef }) content: TemplateRef<any>;
+    @ContentChild(CardTabContentDirective, { read: TemplateRef, static: false }) content: TemplateRef<any>;
 
     constructor(private _tabService: CardTabsService) {
         this._tabService.addTab(this);
