@@ -29,7 +29,7 @@ export class NavigationComponent {
     }
 
     /** Access a custom navigation item template if provided */
-    @ContentChild('uxNavigationItem') navigationItemTemplate: TemplateRef<any>;
+    @ContentChild('uxNavigationItem', { static: false }) navigationItemTemplate: TemplateRef<any>;
 
     /** The classes to be added to each different level */
     _hierarchyClasses = [
